Handle exec and fork in threads (#43)
* do not report on thread clone

we only really care about new pids not new tids

* handle exec* syscalls inside threads

the tid != pid for these execs inside the kprobe but they are the same
inside the kretprobe. This meant that the kretprobe was using a key
different than the kprobe was storing with and thus never completing
the event. We will now store the main thread's pid_tgid() for exec
syscalls so the keys are the same between kprobe and kretprobe

* discard failed unshare + clone-like events