Get real mountpoint (#61)
* cleanup of warning info

use essentially a "thread local" to set/get the warning info

* use the real mountpoint when processing paths

* increase number of path segments to process before tailcalling to 32

* fix argv truncation logic

Our bitwise `&` does a "modulus" rather than a cap. In the other spots
we did this for the buffer length this was fine because the length was
extremely unlikely to cause the "modulus" operator (would only happen
if the buffer became exactly the full length due to a null being
written at the end of the last string). To prevent this we are now
doing a explicit truncation rather than bitwise and