Do not use the buffer to hold the interpreter path
This doesn't work on kernels < 4.18:
https://github.com/torvalds/linux/commit/d71962f3e627b5941804036755c844fabfb65ff5

We are instead going to kee it in the stack. This does push the limits
of our 512 byte stack. llvm-objdump shows we are using 440 bytes of
our stack so we have ~72 bytes of room left for the script_load
program. In the future we may want to split it into different programs
if we need more space since older kernels only allow 128 bytes for the
itnerpreter path so we can reduce our stack size by making a new
program that uses a different constant