fix load_script program to support execveat (#69)
When a program is launched using an fd instead of solely a
path the "filename" the binprm gets is a temporarily allocated string
of the form /dev/fd/*. This string is de-allocated before the the
kretprobe of exec* so relying that "filename" causes us to read
garbagae data (although I don't think it can cause UB since the
probe_read_str function will just fail if it tries to read
uninitialized data). To avoid this, whenever we see a filename that
starts with `/dev/fd/` we will copy the string into our interpreter
path so that we can read it in the kretprobe safely. We will only do
this for the /dev/fd/* case to still optimize for the common case