fix execve* kretprobe for <4.14 kernels (#72)
The issue is that in older kernels bpf_probe_read does not accept
values that could be 0. When we read the argv we didn't check for how
little we were reading, we just checked that we didn't read more than
16KB. To fix this we have added a check for the argv_length being
>1. We are not just checking for >0 because clang converts that into a
!= 0 check and the verifier doesn't use that for bound
tightening. This is OK because non empty strings must be at least two
bytes because of the null byte at the end for C-strings