Initial filemod support monitoring mkdir calls
This contains a lot of relatively minor refactoring to make some code
written for process events also work for file events. Some more
re-organization to move the shared code to a better location is coming.

The file event code is just a starting point, but an event type is in
place and it successfully fills out create events when mkdir is used.