Rearrange code to reflect common code between proc and file events
This is purely re-arranging the existing code by moving .h files around and
splitting the file-events.c into some .h files in a new "file" directory.

Except for the splitting of file-events.c, the only updates should be of paths
in #include directives and the like.

Also, fixed some #import directives to be #include.