Switch mkdir and hardlink to tracepoints (#83)
We should be avoiding using kretprobe of the vfs function calls as
they seem to be called multiple times during overlayfs when
possible. So I switched hardlink to be at the exit of the syscall and
because of that I may as well just put it and mkdir into a tracepoint
to match the other events.

Symlinks are still a special edgecase because of the user string that
we need to copy over but we want to copy the kernel's copy of
it (i.e., what is actually saved) which gets deallocated when the
function ends so we gotta hook before the syscall ends.

Additionally I bumped the minimum number of path segments before tailcalls back up now
that we only process at most one whole path per tailcall.