get data from inode earlier for delete events (#85)
In xfs the inode's mode is cleared during its lazy destruction to mark
the inode as reusable:
https://github.com/torvalds/linux/blob/a0433f8cae3ac51f59b4b1863032822aaa2d8164/fs/xfs/xfs_inode.c#L2298C1-L2298C1. This
means that we cannot grab the mode of the file during the exit of the
syscall since it'd have been cleared already. For safety, I am moving
not just the mode gathering to the security probe but all of the inode
data we care about (mode, ino, device, uid, gid). This means our
internal map takes up a bit more space but that's preferred over
having incorrect information