do not emit KIND_MISMATCH errors on the security probes
these probes can be called by more than just the syscalls that started
the incomplete event. For example, in overlayfs, a modification of a
file in the lower system can cause a new directory being created in
the upper system. This means that an event we saved through FM_MODIFY
will then be attempted to be loaded in a probe that expected FM_CREATE
thus emit a warning and drop the valid event. We will now simply skip
it if the kind isn't what we expected which matches more our old
design of each "kind" being in its own map